+++ CHANGELOG.md
@@ -25,6 +25,7 @@ each for its own `findOne*` or `find*` methods
 * `transaction` method has been removed from `Repository`. Use `EntityManager#transaction` method instead
 * custom repositories do not support container anymore
 * added ActiveRecord support (by extending EntityModel) class
+* controller / subscriber / migrations from options tsconfig now appended with a project root directory
 
 
 ### NEW FEATURES